move shared files from jet/daemon to jet/
