+++ lib/jet/daemon.js
@@ -6,11 +6,12 @@ var EventEmitter = require('events').EventEmitter;
 var MessageSocket = require('./message-socket').MessageSocket;
 
 var jetUtils = require('./utils');
-var Router = require('./daemon/router').Router;
-var jetFetcher = require('./daemon/fetcher');
+var jetFetcher = require('./fetcher');
 var fetchCommon = require('./fetch-common');
+var Elements = require('./element').Elements;
+
 var JsonRPC = require('./daemon/jsonrpc');
-var Elements = require('./daemon/element').Elements;
+var Router = require('./daemon/router').Router;
 var Peers = require('./daemon/peers').Peers;
 
 var isDefined = jetUtils.isDefined;
