+++ lib/jet/element.js
@@ -1,4 +1,4 @@
-var jetUtils = require('../utils');
+var jetUtils = require('./utils');
 
 var Element = function (eachPeerFetcher, owningPeer, path, value) {
 	this.fetchers = {};
