+++ lib/jet/fetch-common.js
@@ -1,6 +1,6 @@
 var jetUtils = require('./utils');
-var jetSorter = require('./daemon/sorter');
-var jetFetcher = require('./daemon/fetcher');
+var jetSorter = require('./sorter');
+var jetFetcher = require('./fetcher');
 
 var checked = jetUtils.checked;
