+++ lib/jet/fetcher.js
@@ -1,7 +1,7 @@
 'use strict';
 var jetPathMatcher = require('./path_matcher');
 var jetValueMatcher = require('./value_matcher');
-var jetUtils = require('../utils');
+var jetUtils = require('./utils');
 
 var isDefined = jetUtils.isDefined;
