+++ lib/jet/peer.js
@@ -3,7 +3,7 @@ var events = require('events');
 var jetUtils = require('./utils');
 var JsonRPC = require('./peer/jsonrpc');
 var fetchCommon = require('./fetch-common');
-var Elements = require('./daemon/element').Elements;
+var Elements = require('./element').Elements;
 
 /**
  * Helpers
