+++ lib/jet/sorter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var jetUtils = require('../utils');
+var jetUtils = require('./utils');
 var isDefined = jetUtils.isDefined;
 
 var createSort = function (options) {
