+++ lib/jet/value_matcher.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var jetUtils = require('../utils');
+var jetUtils = require('./utils');
 
 var generators = {};
